==> analysis_v2/5_FstDynamics.Rmd <==
---
title: "5_AnalyzeFST"
author: "Your Name Here"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(ggpattern)

#setwd("C:/Users/eschlatter/Dropbox/DispersalSGS/analysis")
#treefile = "../output/ts_8238198510834684392_t160000"
#treefile = "../output/638463423/ts_638463423_t500"
#fst_all_file = '../output/500k/FST_all.RData'
#fst_all_file = '../output/FST_all_120k.RData'

#fst_sim_new <- fst_sim
```

Process many single-tick FST files into a single dataframe
```{r, eval=FALSE}
########## Load simulated FSTs, for many timesteps in one simulation

fst_folder_path <- 'C:/Users/eschlatter/Dropbox/DispersalSGS/analysis_v2/full_pop_run/A3'

files <- list.files(path=fst_folder_path) %>% strsplit(" ")
fst_all_sim <- data.frame(sites=factor(),tick=numeric(),fst=numeric(),site1=factor(),site2=factor())

# connect site numbers from simulation to site names (empirical)
sample_sites <- read_csv('../data/sgs_sites_SLiM.csv',show_col_types=FALSE)
sitenames <- data.frame(site_name = sample_sites$Site, site_num = as.factor(1:nrow(sample_sites)))

for (i in 1:length(files)) {
  # load FST matrix
  load(paste0(fst_folder_path,'/',files[[i]])) # load fst matrix
  # extract timestep from filename
  timestep <- strsplit(files[[i]],"_")[[1]][3] %>% substr(start=2, stop=10) %>% as.numeric()
  
  # convert fsts to longer; add column to dataframe
  fst_sim <- as.data.frame(fsts) %>%
    rownames_to_column(var='site1')
  fst_sim <- pivot_longer(fst_sim, cols=2:ncol(fst_sim),names_to=c('site2')) %>%
    mutate(site1=as.factor(site1),site2=as.factor(site2))
  # switch from site numbers to site names
  fst_sim <- left_join(fst_sim,sitenames,by = join_by(site1==site_num)) %>%
    rename(site1_name=site_name) %>%
    left_join(sitenames,by = join_by(site2==site_num)) %>%
    rename(site2_name=site_name) %>%
    dplyr::select(site1=site1_name,site2=site2_name,value)
  fst_sim$sites <- NA
  for(j in 1:nrow(fst_sim)){
    fst_sim$sites[j] <- paste(sort(as.matrix(fst_sim[j,1:2])),collapse='.')
  }
  fst_sim <- filter(fst_sim,!is.na(value)) 
  #%>%rename_with(~paste0('sim_fst_',timestep),value)
  # add column to dataframe
  add_to_df <- fst_sim %>% mutate(tick=timestep) %>% dplyr::select(sites,tick,fst=value,site1,site2)
  fst_all_sim <- rbind(fst_all_sim,add_to_df)
  
}

save(fst_all_sim,file=paste0(fst_folder_path,"/FST_all.RData"))
```

Plot mean FST by tick:
```{r}
runnames <- c("A1","A2","A3","B1","B2","B3")

means=data.frame(tick=seq(200,10600,by=200))

for(i in 1:6){
  fst_folder_path <- paste0('C:/Users/eschlatter/Dropbox/DispersalSGS/analysis_v2/full_pop_run/',runnames[i])
  load(file=paste0(fst_folder_path,"/FST_all.RData"))
  
  ## mean FST by tick
  means_i <- group_by(fst_all_sim,tick) %>%
    summarize(mean_fst = mean(fst))
  colnames(means_i)[2] <- paste0(runnames[i])
  means <- left_join(means,means_i)
}
means <- pivot_longer(means,cols=2:7,names_to="sim",values_to="mean_fst") %>%
  filter(complete.cases(.))

 ggplot(data=means,aes(x=tick))+
    geom_line(aes(y=mean_fst))+
    geom_point(aes(y=mean_fst))+
   facet_wrap(~sim)+
    theme_bw()
 
  ggplot(data=filter(means,sim=="A1"),aes(x=tick))+
    geom_line(aes(y=mean_fst))+
    geom_point(aes(y=mean_fst))+
    theme_bw()
```

Function to plot mean FST by tick, with stability criterion info
```{r}
plot_stability <- function(fst_data, lag_ticks = 500, consecutive_pos_ticks = 1000, tick_interval=50, drop_pts_interval = 1){
  # mean FST by tick
  means <- group_by(fst_data,tick) %>%
    summarize(mean_fst = mean(fst)) %>%
    mutate(mean_lag=mean_fst-lag(mean_fst,n=round(lag_ticks/tick_interval))) # compute the lagged difference
  
  # stability criterion: when the lagged difference doesn't stay positive for more than 1000(?) consecutive ticks
  # 1=unstable, 0=stable
  num_consec <- consecutive_pos_ticks/tick_interval
  stable_crit=c()
  for(i in 1:nrow(means)){ # if there haven't been num_consec values yet, we can't evaluate stability
    if(i<num_consec){
      stable_crit[i]=NA
    }
    else { # if there have been enough values already, check them all
      is_stable = prod(means$mean_lag[(i-num_consec+1):i]>0)
      stable_crit[i]=is_stable
    }
  }
  means$stable_crit <- as.factor(stable_crit)
  
  #stability threshold: the last time the series was unstable
  stable_tick <- means$tick[max(which(stable_crit==1))]

  # visualize it  
  ggplot(data=means[seq(from=1,to=nrow(means),by=drop_pts_interval),],aes(x=tick))+
    geom_line(aes(y=mean_fst))+
    geom_point(aes(y=mean_fst, color=stable_crit))+
    geom_vline(xintercept=stable_tick)+
    scale_color_manual(labels=c("stable","unstable"),values=c("cyan3","brown1"))+
    annotate("text",x=stable_tick+25,y=min(means$mean_fst),label=paste0(stable_tick), hjust=0)+
    labs(title=paste0("Stability of FST: lag = ",lag_ticks,", consecutive positive = ",consecutive_pos_ticks))+
    theme_minimal()
}
```

## Reduced-population-size test runs

```{r}
fst_folder_path <- 'C:/Users/eschlatter/Dropbox/DispersalSGS/analysis_v2/test_fst_longer'
load(file=paste0(fst_folder_path,"/FST_all.RData"))

plot_stability(fst_all_sim, lag_ticks = 500, consecutive_pos_ticks = 1000, tick_interval=100)
#plot_stability(fst_all_sim, lag_ticks = 500, consecutive_pos_ticks = 750, tick_interval=100)
#plot_stability(fst_all_sim, lag_ticks = 750, consecutive_pos_ticks = 500,tick_interval=100)
```

```{r}
fst_folder_path <- 'C:/Users/eschlatter/Dropbox/DispersalSGS/analysis_v2/test_fst_longer_again'
load(file=paste0(fst_folder_path,"/FST_all_more.RData"))

plot_stability(fst_all_sim, lag_ticks = 500, consecutive_pos_ticks = 1000, tick_interval=50)
#plot_stability(fst_all_sim, lag_ticks = 500, consecutive_pos_ticks = 750, tick_interval=50)
#plot_stability(fst_all_sim, lag_ticks = 750, consecutive_pos_ticks = 500, tick_interval=50)
```


## Full-population-size test runs


```{r}
fst_folder_path <- 'C:/Users/eschlatter/Dropbox/DispersalSGS/analysis_v2/500k_05km_1'
load(file=paste0(fst_folder_path,"/FST_all_more.RData"))

## mean FST by tick
means <- group_by(fst_all_sim,tick) %>%
  summarize(mean_fst = mean(fst))

ggplot(data=means,aes(x=tick))+
  geom_line(aes(y=mean_fst))+
  geom_point(aes(y=mean_fst))+
  theme_minimal()
```

```{r}
fst_folder_path <- 'C:/Users/eschlatter/Dropbox/DispersalSGS/analysis_v2/500k_05km_2'
load(file=paste0(fst_folder_path,"/FST_all.RData"))

## mean FST by tick
means <- group_by(fst_all_sim,tick) %>%
  summarize(mean_fst = mean(fst))

ggplot(data=means,aes(x=tick))+
  geom_line(aes(y=mean_fst))+
  geom_point(aes(y=mean_fst))+
  theme_minimal()
```





















Once we have the pairwise FST values by tick, let's do something with them:
```{r}
fst_folder_path <- 'C:/Users/eschlatter/Dropbox/DispersalSGS/analysis_v2/test_fst_longer'
load(file=paste0(fst_folder_path,"/FST_all.RData"))

## mean FST by tick
means <- group_by(fst_all_sim,tick) %>%
  summarize(mean_fst = mean(fst))
plot(means$tick,means$mean_fst, type='l')


## stop criteria
means <- mutate(means,
                delta_1=mean_fst-lag(mean_fst),
                delta_2=mean_fst-lag(mean_fst,n=2),
                delta_3=mean_fst-lag(mean_fst,n=3),
                delta_4=mean_fst-lag(mean_fst,n=4),
                delta_5=mean_fst-lag(mean_fst,n=5))

ggplot(data=means[seq(from=1,to=99,by=1),],aes(x=tick))+
  geom_line(aes(y=mean_fst))+
  geom_point(aes(y=mean_fst))+
  #geom_line(aes(y=delta_1), color="red")+
  geom_line(aes(y=delta_5), color="black")

# criterion: when delta_5 doesn't stay positive for more than 1000(?) consecutive ticks
num_consec <- 10
stable_crit=c()
for(i in 1:nrow(means)){ # if there haven't been 10 values yet, we can't evaluate stability
  if(i<num_consec){
    stable_crit[i]=NA
  }
  else { # if there have been 10 values already, check them all
    is_stable = prod(means$delta_5[(i-num_consec+1):i]>0)
    stable_crit[i]=is_stable
  }
}
means$stable_crit <- as.factor(stable_crit)
stable_tick <- means$tick[min(which(stable_crit==0))]

ggplot(data=means[seq(from=1,to=99,by=1),],aes(x=tick))+
  geom_line(aes(y=mean_fst))+
  geom_point(aes(y=mean_fst, color=stable_crit))+
  #geom_line(aes(y=delta_1), color="red")+
  #geom_line(aes(y=delta_5), color="black")+
  geom_vline(xintercept=stable_tick)+
  scale_color_manual(labels=c("stable","unstable"),values=c("cyan3","brown1"))+
  theme_minimal()
```

```{r}
fst_folder_path <- 'C:/Users/eschlatter/Dropbox/DispersalSGS/analysis_v2/test_fst_longer_again'
load(file=paste0(fst_folder_path,"/FST_all_more.RData"))

## mean FST by tick
means <- group_by(fst_all_sim,tick) %>%
  summarize(mean_fst = mean(fst))

ggplot(data=means,aes(x=tick))+
  geom_line(aes(y=mean_fst))+
  geom_point(aes(y=mean_fst))+
  theme_minimal()

## stop criteria
means <- mutate(means,
                delta_1=mean_fst-lag(mean_fst),
                delta_2=mean_fst-lag(mean_fst,n=2),
                delta_3=mean_fst-lag(mean_fst,n=3),
                delta_4=mean_fst-lag(mean_fst,n=4),
                delta_5=mean_fst-lag(mean_fst,n=10))

ggplot(data=means,aes(x=tick))+
  geom_line(aes(y=mean_fst))+
  geom_point(aes(y=mean_fst))+
  #geom_line(aes(y=delta_1), color="red")+
  geom_line(aes(y=delta_5), color="black")

# criterion: when delta_5 doesn't stay positive for more than 1000(?) consecutive ticks
num_consec <- 15
stable_crit=c()
for(i in 1:nrow(means)){ # if there haven't been 10 values yet, we can't evaluate stability
  if(i<num_consec){
    stable_crit[i]=NA
  }
  else { # if there have been 10 values already, check them all
    is_stable = prod(means$delta_5[(i-num_consec+1):i]>0)
    stable_crit[i]=is_stable
  }
}
means$stable_crit <- as.factor(stable_crit)
stable_tick <- means$tick[max(which(stable_crit==1))]

ggplot(data=means,aes(x=tick))+
  geom_line(aes(y=mean_fst))+
  geom_point(aes(y=mean_fst, color=stable_crit))+
  #geom_line(aes(y=delta_1), color="red")+
  #geom_line(aes(y=delta_5), color="black")+
  geom_vline(xintercept=stable_tick)+
  scale_color_manual(labels=c("stable","unstable"),values=c("cyan3","brown1"))+
  theme_minimal()

#https://ggplot2-book.org/annotations#sec-custom-annotations
```
Function to plot mean FST by tick, with stability criterion info
```{r}
plot_stability <- function(fst_data, lag_ticks = 500, consecutive_pos_ticks = 1000, tick_interval=50){
  # mean FST by tick
  means <- group_by(fst_data,tick) %>%
    summarize(mean_fst = mean(fst)) %>%
    mutate(mean_lag=mean_fst-lag(mean_fst,n=round(lag_ticks/tick_interval))) # compute the lagged difference
  
  # stability criterion: when the lagged difference doesn't stay positive for more than 1000(?) consecutive ticks
  # 1=unstable, 0=stable
  num_consec <- consecutive_pos_ticks/tick_interval
  stable_crit=c()
  for(i in 1:nrow(means)){ # if there haven't been num_consec values yet, we can't evaluate stability
    if(i<num_consec){
      stable_crit[i]=NA
    }
    else { # if there have been enough values already, check them all
      is_stable = prod(means$mean_lag[(i-num_consec+1):i]>0)
      stable_crit[i]=is_stable
    }
  }
  means$stable_crit <- as.factor(stable_crit)
  
  #stability threshold: the last time the series was unstable
  stable_tick <- means$tick[max(which(stable_crit==1))]

  # visualize it  
  ggplot(data=means,aes(x=tick))+
    geom_line(aes(y=mean_fst))+
    geom_point(aes(y=mean_fst, color=stable_crit))+
    geom_vline(xintercept=stable_tick)+
    scale_color_manual(labels=c("stable","unstable"),values=c("cyan3","brown1"))+
    annotate("text",x=stable_tick+25,y=min(means$mean_fst),label=paste0(stable_tick), hjust=0)+
    labs(title=paste0("Stability of FST: lag = ",lag_ticks,", consecutive positive = ",consecutive_pos_ticks))+
    theme_minimal()
}
```


```{r}
fst_folder_path <- 'C:/Users/eschlatter/Dropbox/DispersalSGS/analysis_v2/test_fst_longer_again'
load(file=paste0(fst_folder_path,"/FST_all_more.RData"))

plot_stability(fst_all_sim, lag_ticks = 500, consecutive_pos_ticks = 1000)
plot_stability(fst_all_sim, lag_ticks = 500, consecutive_pos_ticks = 750)
plot_stability(fst_all_sim, lag_ticks = 750, consecutive_pos_ticks = 500)
```
```{r}
fst_folder_path <- 'C:/Users/eschlatter/Dropbox/DispersalSGS/analysis_v2/test_fst_longer'
load(file=paste0(fst_folder_path,"/FST_all.RData"))

plot_stability(fst_all_sim, lag_ticks = 500, consecutive_pos_ticks = 1000, tick_interval=100)
plot_stability(fst_all_sim, lag_ticks = 500, consecutive_pos_ticks = 750, tick_interval=100)
plot_stability(fst_all_sim, lag_ticks = 750, consecutive_pos_ticks = 500,tick_interval=100)
```






































```{r}
########## Combine simulated and empirical

# combine simulation runs into a long table, with a column for the simulation number
all_fst_sim <- data.frame(site1=factor(), site2=factor(), sim_fst=numeric(), sites=factor(), sim=factor())
for(i in 1:length(fst_sims)){
  this_sim <- fst_sims[[i]]
  all_fst_sim <- rbind(all_fst_sim,mutate(this_sim, sim=as.factor(i)))
}

# add empirical data as its own column
all_fst_both <- full_join(all_fst_sim,fst_emp,by='sites') %>%
  select(sites,site1,site2=site2.x,sim,sim_fst,emp_fst) %>%
  filter(!is.na(emp_fst))

########## Group by site pair (across simulations)

all_fst_group <- group_by(all_fst_both,sites) %>%
  summarize(emp_fst=first(emp_fst),site1=first(site1),site2=first(site2),
            sim_med=median(sim_fst),sim_sd=sd(sim_fst),sim_min=min(sim_fst),sim_max=max(sim_fst))

# add extra columns for site, population, and population pair, to make exploratory plotting easier
site_list <- unique(c(fst_emp$X,fst_emp$site2))
site_list <- c(site_list[length(site_list)],site_list[1:(length(site_list)-1)]) # put them in order
for(i in site_list){
  all_fst_group <- mutate(all_fst_group,newsite=(i==site1 | i==site2)) %>%
    rename_with(~paste0(i),newsite)
}
for(i in c('B','T','G','L')){
  all_fst_group <- mutate(all_fst_group,newsite=grepl(i,sites)) %>%
    rename_with(~paste0(i),newsite)
}
all_fst_group <- mutate(all_fst_group,first_pop=substr(sites,1,1)) %>%
  mutate(second_pop=substr(sub('.*\\.','',sites),1,1)) %>%
  mutate(both_pops=paste0(first_pop,second_pop))

save(all_fst_group,file=fst_all_file)
```

Get pairwise distance between sites (only need to do this once)
```{r, eval=FALSE}
load(fst_all_file)
sample_sites <- read_csv('../data/sgs_sites_SLiM',show_col_types=FALSE)

sitepair_dists <- dplyr::select(all_fst_group,sites,site1,site2)
sitepair_dists <- left_join(sitepair_dists,sample_sites,by=c('site1' ='Site'),keep=FALSE) %>%
  select(sites,site1,site2,site1.x=x,site1.y=y) %>%
  left_join(sample_sites,by=c('site2' ='Site'),keep=FALSE) %>%
  select(sites,site1,site2,site1.x,site1.y,site2.x=x,site2.y=y) %>%
  mutate(dist=sqrt((site1.x-site2.x)^2+(site1.y-site2.y)^2))

save(sitepair_dists, file='../data/sitepair_dists.RData')

```

```{r}
load(fst_all_file)

all_fst_group <- mutate(all_fst_group,both_pops_bt=case_match(both_pops,'BT'~'BB','GT'~'BG','LT'~'BL','TT'~'BB',.default=both_pops))
all_fst_group$both_pops_bt <- factor(all_fst_group$both_pops_bt,levels=c('BB','GG','LL','BG','BL','GL'))

#'#fc8d59' -orange (G)
#'#ffffbf' -yellow (L)
#'#91bfdb' -blue (B)

ggplot(data=all_fst_group,aes(x=sim_med,fill=both_pops_bt,pattern_fill=both_pops_bt,pattern_density=both_pops_bt))+
  geom_density_pattern(pattern_spacing=0.05,alpha=0.6,pattern_alpha=0.6)+
  scale_pattern_density_manual(values=c(0,0,0,0.5,0.5,0.5),name='Populations')+
  scale_fill_manual(values=c('#91bfdb','#fc8d59','#ffffbf','#91bfdb','#91bfdb','#fc8d59'),name='Populations')+
  scale_pattern_fill_manual(values=c('#91bfdb','#fc8d59','#ffffbf','#fc8d59','#ffffbf','#ffffbf'),name='Populations')+
  theme_bw()+
    theme(legend.position='bottom')+
  xlab('Simulated FST')+
  guides(pattern_density = guide_legend(nrow = 3))

ggplot(data=all_fst_group,aes(x=sim_med,fill=both_pops_bt,pattern_fill=both_pops_bt,pattern_density=both_pops_bt))+
  geom_density_pattern(pattern_spacing=0.03,alpha=0.6,pattern_alpha=0.6)+
  scale_pattern_density_manual(values=c(0,0,0,0.5,0.5,0.5),name='Populations')+
  scale_fill_manual(values=c('#91bfdb','#fc8d59','#ffffbf','#91bfdb','#91bfdb','#fc8d59'),name='Populations')+
  scale_pattern_fill_manual(values=c('#91bfdb','#fc8d59','#ffffbf','#fc8d59','#ffffbf','#ffffbf'),name='Populations')+
  theme_bw()+
  theme(legend.position='none', panel.grid.major = element_blank(),panel.grid.minor=element_blank())+
  labs(title='Simulated',x=NULL,y=NULL)

ggplot(data=all_fst_group,aes(x=emp_fst,fill=both_pops_bt,pattern_fill=both_pops_bt,pattern_density=both_pops_bt))+
  geom_density_pattern(pattern_spacing=0.03,alpha=0.6,pattern_alpha=0.6)+
  scale_pattern_density_manual(values=c(0,0,0,0.5,0.5,0.5),name='Populations')+
  scale_fill_manual(values=c('#91bfdb','#fc8d59','#ffffbf','#91bfdb','#91bfdb','#fc8d59'),name='Populations')+
  scale_pattern_fill_manual(values=c('#91bfdb','#fc8d59','#ffffbf','#fc8d59','#ffffbf','#ffffbf'),name='Populations')+
  theme_bw()+
  theme(legend.position='none', panel.grid.major = element_blank(),panel.grid.minor=element_blank())+
  labs(title='Empirical',x=NULL,y=NULL)
```

```{r,warning=FALSE}
plotlist <- list()
mse <- vector()
mean_fst <- vector()
v_t <- c(20,60,120,180,240,300)

for(i in 1:length(v_t)){
  t <- v_t[i]
  fst_all_file = paste0('../output/FST_all_',t,'k.RData')
  load(fst_all_file)
  all_fst_group <- mutate(all_fst_group,both_pops_bt=case_match(both_pops,'BT'~'BB','GT'~'BG','LT'~'BL','TT'~'BB',.default=both_pops))
  all_fst_group$both_pops_bt <- factor(all_fst_group$both_pops_bt,levels=c('BB','GG','LL','BG','BL','GL'))
  
  g <- ggplot(data=all_fst_group,aes(x=sim_med,fill=both_pops_bt,pattern_fill=both_pops_bt,pattern_density=both_pops_bt))+
    geom_density_pattern(pattern_spacing=0.03,alpha=0.6,pattern_alpha=0.6)+
    scale_pattern_density_manual(values=c(0,0,0,0.5,0.5,0.5),name='Populations')+
    scale_fill_manual(values=c('#91bfdb','#fc8d59','#ffffbf','#91bfdb','#91bfdb','#fc8d59'),name='Populations')+
    scale_pattern_fill_manual(values=c('#91bfdb','#fc8d59','#ffffbf','#fc8d59','#ffffbf','#ffffbf'),name='Populations')+
    theme_bw()+
    xlim(0,0.25)+
    ylim(0,450)+
    theme(legend.position='none', panel.grid.major = element_blank(),panel.grid.minor=element_blank())+
    labs(title=paste0('t = ',t),x=NULL,y=NULL)
  
  print(g)
  plotlist[[i]] <- g
  
  mse[i] <- mean((all_fst_group$sim_med - all_fst_group$emp_fst)^2)
  mean_fst[i] <- mean(all_fst_group$sim_med)
}

library(patchwork)
#h <- wrap_plots(plotlist,nrow=2)
```


Again, but let's do subsets
```{r}
load(fst_all_file)
sample_sites <- read_csv('../data/sgs_sites_SLiM',show_col_types=FALSE)

fsts <- dplyr::select(all_fst_group,sites,site1,site2,emp_fst,sim_med) %>%
#  filter(!(site1 %in%c('B1','B2','B3'))&!(site2 %in%c('B1','B2','B3'))) %>%
  left_join(sample_sites,by=c('site1'='Site')) %>%
  dplyr::select(sites,site1,site2,emp_fst,sim_med,region1=Region) %>%
  left_join(sample_sites,by=c('site2'='Site')) %>%
  dplyr::select(sites,site1,site2,emp_fst,sim_med,region1,region2=Region) %>%
  mutate(region1=case_match(region1,
                            'barrier (northern)'~'N','barrier (central)'~'C',
                            'Turneffe Atoll'~'T',"Glover's Atoll" ~ 'G','Lighthouse Atoll'~'L',
                            .default='G')) %>%
  mutate(region2=case_match(region2,
                            'barrier (northern)'~'N','barrier (central)'~'C',
                            'Turneffe Atoll'~'T',"Glover's Atoll" ~ 'G','Lighthouse Atoll'~'L',
                            .default='G'))

fsts_subset <- filter(fsts,(region1 %in% c('N','C','T'))&(region2 %in% c('N','C','T')))

fsts_subset$regions <- NA
for(j in 1:nrow(fsts_subset)){
  fsts_subset$regions[j] <- paste(sort(as.matrix(fsts_subset[j,c('region1','region2')])),collapse='')
  }

fsts_subset$regions <- factor(fsts_subset$regions,levels=c('NN','CC','TT','CN','NT','CT'))

#'#2b8cbe' -northern (N)
#'#a6bddb' -central (C)
#'#ece7f2' -turneffe (T)

ggplot(data=fsts_subset,aes(x=sim_med,fill=regions,pattern_fill=regions,pattern_density=regions))+
  geom_density_pattern(pattern_spacing=0.05,alpha=0.5,pattern_alpha=0.5)+
  scale_pattern_density_manual(values=c(0,0,0,0.5,0.5,0.5),name='Regions')+
  scale_fill_manual(values=c('#2b8cbe','#a6bddb','#ece7f2','#2b8cbe','#2b8cbe','#a6bddb'),name='Regions')+
  scale_pattern_fill_manual(values=c('#2b8cbe','#a6bddb','#ece7f2','#a6bddb','#ece7f2','#ece7f2'),name='Regions')+
  theme_bw()+
  xlab('Simulated FST')+
    theme_bw()+
    theme(legend.position='bottom')+
  xlab('Simulated FST')+
  guides(pattern_density = guide_legend(nrow = 3))

ggplot(data=fsts_subset,aes(x=sim_med,fill=regions,pattern_fill=regions,pattern_density=regions))+
  geom_density_pattern(pattern_spacing=0.02,alpha=0.5,pattern_alpha=0.5)+
  scale_pattern_density_manual(values=c(0,0,0,0.5,0.5,0.5),name='Regions')+
  scale_fill_manual(values=c('#2b8cbe','#a6bddb','#ece7f2','#2b8cbe','#2b8cbe','#a6bddb'),name='Regions')+
  scale_pattern_fill_manual(values=c('#2b8cbe','#a6bddb','#ece7f2','#a6bddb','#ece7f2','#ece7f2'),name='Regions')+
  theme_bw()+
  theme(legend.position='none', panel.grid.major = element_blank(),panel.grid.minor=element_blank())+
  labs(title='Simulated',x=NULL,y=NULL)

ggplot(data=fsts_subset,aes(x=emp_fst,fill=regions,pattern_fill=regions,pattern_density=regions))+
  geom_density_pattern(bw=0.0005,pattern_spacing=0.02,alpha=0.6,pattern_alpha=0.6)+ # try bw=0.0005 to remove lumpiness
  scale_pattern_density_manual(values=c(0,0,0,0.5,0.5,0.5),name='Regions')+
  scale_fill_manual(values=c('#2b8cbe','#a6bddb','#ece7f2','#2b8cbe','#2b8cbe','#a6bddb'),name='Regions')+
  scale_pattern_fill_manual(values=c('#2b8cbe','#a6bddb','#ece7f2','#a6bddb','#ece7f2','#ece7f2'),name='Regions')+
  theme_bw()+
  ylim(0,1000)+
  theme(legend.position='none', panel.grid.major = element_blank(),panel.grid.minor=element_blank())+
  labs(title='Empirical',x=NULL,y=NULL)

```

Check IBD
```{r}
load(fst_all_file)
load('../data/sitepair_dists.RData')

all_fst_group <- left_join(all_fst_group,sitepair_dists,keep=FALSE) %>%
  filter(both_pops=='BB')

all_fst_group <- mutate(all_fst_group,
                        emp_scale=emp_fst/(1-emp_fst),
                        sim_scale=sim_med/(1-sim_med))

ggplot(all_fst_group,aes(x=dist,y=sim_scale))+
  geom_point(size=1.1)+
  theme_bw()+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  labs(title='Simulated',x=NULL,y=NULL)
#  geom_smooth(method='lm',se=FALSE)+


ggplot(all_fst_group,aes(x=dist,y=sim_med))+
  geom_point(size=1.1)+
  theme_bw()+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  labs(title='Simulated',x=NULL,y=NULL)
#  geom_smooth(method='lm',se=FALSE)+


ggplot(all_fst_group,aes(x=dist,y=emp_scale))+
  geom_point(size=1.1)+
  theme_bw()+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  labs(title='Simulated',x=NULL,y=NULL)
#  geom_smooth(method='lm',se=FALSE)+

ggplot(all_fst_group,aes(x=dist,y=emp_fst))+
  geom_point(size=1.1)+
  theme_bw()+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  labs(title='Empirical',x=NULL,y=NULL)
```
